python-projects/api-dev/azure-apis/kv-api/requirements.txt
Flask==3.0.0
azure-identity==1.15.0
azure-mgmt-keyvault==10.3.0
azure-mgmt-resource==23.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
